+++ quisp/rules/actions/SwappingAction_test.cc
@@ -40,7 +40,7 @@ class SwappingAction : public OriginalSwappingAction {
   using OriginalSwappingAction::SwappingAction;
   static std::unique_ptr<SwappingAction> setupAction() {
     unsigned long ruleset_id = 0;
-    int rule_index = 1;
+    unsigned long rule_index = 1;
 
     int left_partner_addr = 2;
     QNIC_type left_qnic_type = QNIC_E;
